+++ server/src/models/post.js
@@ -45,8 +45,8 @@ module.exports = (sequelize, DataTypes) => {
       userId: DataTypes.STRING,
       overviewId: DataTypes.STRING,
       imagesId: DataTypes.STRING,
-      priceNumber: DataTypes.INTEGER,
-      areaNumber: DataTypes.INTEGER,
+      priceNumber: DataTypes.FLOAT,
+      areaNumber: DataTypes.FLOAT,
       createdAt: DataTypes.DATE,
     },
     {
